Bugfix - to reset new employee form
The state was being recalled if an employee's file was viewed before the new employee form was selected. A reset state has been added for when the New Employee form is accessed. This was omitted from the course but a couple of guys picked up on it in the Q&A for lecture 156.

Text was offset due to an error in styling.